Move code into layered model architecture
