+++ .env
@@ -1,0 +1,2 @@
+HYPERNOVA_BATCH=http://hypernova:3000/batch
+CONFIG_FILE=config.json
